requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
redis==5.1.0
pydantic==2.9.2
orjson>=3.9.0
langsmith>=0.1.68
langchain>=0.2.10
langchain-community>=0.2.10
langchain-openai>=0.1.23
openai>=1.35.10